==> research/02_data_validation.ipynb <==
import os
os.chdir('../')

from dataclasses import dataclass
from pathlib import Path

@dataclass
class DataValidatioConfig:
   root_dir: Path
   status_file: Path
   ALL_REQUIRED_FILES: list

from titanicSurvival.constants import *
from titanicSurvival.utils.common import  read_yaml, create_directories

class ConfigurationManager:
    def __init__(self,
                 config_file_path= CONFIG_FILE_PATH,
                 params_file_path= PARAMS_FILE_PATH):
        self.config = read_yaml(config_file_path)
        self.params = read_yaml(params_file_path)

        create_directories([self.config.artifacts_root])
    def get_data_validation(self) -> DataValidatioConfig:
        config = self.config.data_validation

        create_directories([config.root_dir])

        data_valiation_config = DataValidatioConfig(
            root_dir=config.root_dir,
            status_file=config.status_file,
            ALL_REQUIRED_FILES=config.ALL_REQUIRED_FILES
        )
        return data_valiation_config
    


import os
import urllib.request as request
import zipfile
from titanicSurvival.logging import logger
from titanicSurvival.utils.common import get_size


class DataValidation:
    def __init__(self,config:DataValidatioConfig):
        self.config =config
    
    def validate_files_exist(self) -> bool:
        try:
            for file in self.config.ALL_REQUIRED_FILES:
                if os.path.exists(os.path.join('artifacts/data_ingestion',file)):
                    with open(self.config.status_file, 'w') as fp:
                        fp.write(f'File exising: {file}')
                    logger.info(f"File Existing: {file}")
                    validation_status = True
                else:
                    with open(self.config.status_file, 'w') as fp:
                        fp.write(f'File Not exising: {file}')
                    validation_status = False

                    logger.info(f"File Not Existing: {file}")
            return validation_status
        except Exception  as e:
            raise e

try:
    config = ConfigurationManager()
    data_validation_config = config.get_data_validation()
    datavalidation = DataValidation(config=data_validation_config)
    datavalidation.validate_files_exist()
    
except Exception as e:
    raise e
