==> research/04_model_trainer.ipynb <==
import os
os.chdir('../')

%pwd

from dataclasses import dataclass
from pathlib import Path

@dataclass
class ModelTrainerConfig:
    root_dir: Path
    data_path: Path

from titanicSurvival.constants import *
from titanicSurvival.utils.common import  read_yaml, create_directories

class ConfigurationManager:
    def __init__(self,
                 config_file_path= CONFIG_FILE_PATH,
                 params_file_path= PARAMS_FILE_PATH):
        self.config = read_yaml(config_file_path)
        self.params = read_yaml(params_file_path)

        create_directories([self.config.artifacts_root])
    def get_model_trainer(self) -> ModelTrainerConfig:
        config = self.config.model_trainer

        create_directories([config.root_dir])

        model_trainer_config = ModelTrainerConfig(
            root_dir=config.root_dir,
            data_path=config.data_path
        )
        return model_trainer_config

import os
import urllib.request as request
import zipfile
from titanicSurvival.logging import logger
from titanicSurvival.utils.common import get_size
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.preprocessing import OrdinalEncoder
from sklearn.metrics import classification_report
from typing import Any

from sklearn.linear_model import LogisticRegression

import joblib

class ModelTrainer:
    def __init__(self,config:ModelTrainerConfig):
        self.config =config
    
    def get_transformed_data(self) -> tuple[np.array, np.array, np.array, np.array]:
        try:
            X_train_df = pd.read_csv(os.path.join('artifacts/data_transformation/','train_features_x.csv'))
            y_train_df = pd.read_csv(os.path.join('artifacts/data_transformation/','train_features_y.csv'))
            X_test_df = pd.read_csv(os.path.join('artifacts/data_transformation/','test_features_x.csv'))
            y_test_df = pd.read_csv(os.path.join('artifacts/data_transformation/','test_features_y.csv'))
            
            logger.info(f"Cocatenated train and test data")
            print(X_train_df.head())
            return X_train_df.to_numpy(),y_train_df.to_numpy(),X_test_df.to_numpy(),y_test_df.to_numpy()
        except Exception as e:
            raise e
    
    
    def fit_model(self,X_train,y_train,X_test,y_test):

        X_train, y_train, X_test, y_test = self.get_transformed_data()

        lr = LogisticRegression()

        lr.fit(X_train,y_train)
        logger.info(f"{ classification_report(y_test,lr.predict(X_test))}")

        joblib.dump(lr,os.path.join(self.config.root_dir,'model.pkl'))
        logger.info(f"Model Save to path: {self.config.root_dir}")

            


# try:
#     config = ConfigurationManager()
#     modletrainer_config = config.get_model_trainer()
#     model_trainer=ModelTrainer(config=modletrainer_config)
#     X_train,y_train,X_test,y_test =model_trainer.get_transformed_data()
#     model_trainer.fit_model(X_train,y_train,X_test,y_test)
    
# except Exception as e:
#     raise e

from titanicSurvival.components.data_transformation import assignCabin, getAgeSubSection

def convert_input_data_exp_feature_vector(input_data):
    df  = pd.DataFrame(input_data)
    df['Sex'] = df['Sex'].fillna(df['Sex'].mode()[0]).map({'female':0, 'male': 1 })
    df['Age']=df['Age'].fillna(df['Age'].mode()[0])
    df['Age']=df['Age'].apply(lambda x: getAgeSubSection(x))
    df['Embarked']=df['Embarked'].fillna(df['Embarked'].mode()[0])
    df['Embarked'] = (df['Embarked'].map({'C':0, 'S':1, 'Q':2})).astype(int)
    
    df['Cabin']=df['Cabin'].fillna("X").map(lambda x: x[0])
    df['Cabin']=df.apply(lambda row: assignCabin(row),axis=1)
    print(f"Cabin data after assigned values: {df['Cabin'].head()}")

    cabinEncode = OrdinalEncoder()
    df['Cabin']=(cabinEncode.fit_transform(df[['Cabin']])).astype(int)
    df.drop(columns=['PassengerId','Name','Ticket','Fare'],inplace=True,axis=1)
    #print(df['Cabin'].isnull().sum())
    print(df.head())
    df_np = df.to_numpy()
    return df_np

from io import StringIO
import pandas as pd
df_eval = pd.DataFrame(StringIO("""PassengerId,Pclass,Name,Sex,Age,SibSp,Parch,Ticket,Fare,Cabin,Embarked
892,3,"Kelly, Mr. James",male,34.5,0,0,330911,7.8292,,Q"""))

import pandas as pd

data = {
    "PassengerId": [893],
    "Pclass": [3],
    "Name": ["Wilkes, Mrs. James (Ellen Needs)"],
    "Sex": ["female"],
    "Age": [47],
    "SibSp": [1],
    "Parch": [0],
    "Ticket": ["363272"],
    "Fare": [7],
    "Cabin": [None],  # Missing value
    "Embarked": ["S"]
}

df_eval = pd.DataFrame(data)
print(df_eval)


df_eval = pd.read_csv('./artifacts/data_ingestion/test.csv')
df_eval

y_target = pd.read_csv('./artifacts/data_ingestion/gender_submission.csv')
y_target.drop(columns=['PassengerId'],inplace=True,axis=1)
y_target.head(2)
y_target_np = y_target.to_numpy()

try:
    config = ConfigurationManager()
    modletrainer_config = config.get_model_trainer()
    features = convert_input_data_exp_feature_vector(df_eval)
    model_load = joblib.load(os.path.join(modletrainer_config.root_dir,'model.pkl'))
    predicted_out =model_load.predict(features)

    print(classification_report(y_target_np[:10],predicted_out[:10] ,target_names=['Not Survived','Survived']))
    #print(features)
except Exception as e:
    raise e
